pyproject.toml
[project]
name = "ccme"
version = "0.1.0"
requires-python = ">=3.10.6"
dependencies = [
  "chex>=0.1.90",
  "equinox",
  "hydra-core>=1.3.2",
  "jax[cuda12]",
  "matplotlib>=3.10.1",
  "numpy>=2.2.5",
  "optax>=0.2.4",
  "orbax>=0.1.9",
  "pandas>=2.3.3",
  "scikit-learn>=1.6.1",
  "scipy>=1.15.2",
  "seaborn>=0.13.2",
  "six>=1.17.0",
  "tqdm>=4.67.1",
]
